backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
httpx>=0.24.0
email-validator>=2.0.0
